backend/requirements.txt
# FastAPI Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.12.3
pydantic-settings==2.7.1

# Database
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Redis
redis==5.2.1
hiredis==3.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
bcrypt==4.2.1

# Astrology Library
kerykeion==5.0.2

# HTTP Client
httpx==0.28.1
requests==2.32.4

# Date/Time
pytz==2025.2
python-dateutil==2.9.0.post0

# Internationalization
babel==2.16.0

# File Storage
boto3==1.40.57  # AWS S3 (optional)

# Validation & Utils
email-validator==2.2.0
phonenumbers==8.13.54

# Logging
python-json-logger==4.0.0

# Environment
python-dotenv==1.0.1

# CORS - FastAPI includes CORS middleware, no extra package needed
